Fix to avoid Signal bar display
Voice & Data is in service over IWLAN and cellular Voice is OOS &
Data is in limited service. since Data is in limited service over
Cellular network, modem reports signal strength and Voice is in
service over IWLAN,signal strength is displayed. additionally
check whether Voice RAT is valid or not to report hasService.

Change-Id: Ib4d569237624870b054b7b62efa8d886bcaa8f23