Fix for AccountSelection Dialog in Lockscreen issue
Phone Account Selection Dialog is displayed over lockscreen for
an outgoing call. But since the dialog gets onStop from framework,
the dialog and hence the call gets disconnected. This is due to
delay in sleeping state in framework as per current lock screen
design

Current Fix tries to determine the sleeping state not just based
on the state variable but also on the wakefulness state of the
device to determine sleeping state.
CRs-Fixed: 1082943

Change-Id: Ideaf128de9f722e58b107efda020350a0ead0304